Moves some more relevant pages to variant
