+++ scout/server/blueprints/variant/templates/variant/variant.html
@@ -328,7 +328,7 @@
           </ul>
         </div>
       </ul>
-      <form action="{{ url_for('variants.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
+      <form action="{{ url_for('variant.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
         <label class="mt-3">Variant tag</label>
         <div class="row">
           <div class="col-8">
@@ -346,7 +346,7 @@
           </div>
         </div>
       </form>
-      <form action="{{ url_for('variants.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
+      <form action="{{ url_for('variant.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
         <label class="mt-3">Dismiss variant</label>
         <div class="row">
           <div class="col-8">
@@ -363,7 +363,7 @@
           </div>
         </div>
       </form>
-      <form action="{{ url_for('variants.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
+      <form action="{{ url_for('variant.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
         <label class="mt-3">Mosaic Tags</label>
         <div class="row">
           <div class="col-8">
@@ -382,9 +382,9 @@
       </form>
       {{ acmg_form(variant.acmg_classification.code if variant.acmg_classification) }}
       <div class="mt-3">
-        <a href="{{ url_for('variants.variant_acmg', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" class="btn btn-outline-secondary form-control">Classify</a>
+        <a href="{{ url_for('variant.variant_acmg', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" class="btn btn-outline-secondary form-control">Classify</a>
         {% if variant._id in case.suspects and not variant.clinvar_clinsig %}
-          <a href="{{ url_for('variants.clinvar', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" class="btn btn-outline-secondary form-control">Submit to ClinVar</a>
+          <a href="{{ url_for('variant.clinvar', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" class="btn btn-outline-secondary form-control">Submit to ClinVar</a>
         {% endif %}
         {% if variant.clinvar_clinsig %}
           <a href="{{ url_for('cases.clinvar_submissions', institute_id=institute._id) }}" class="btn btn-outline-secondary form-control">Modify clinvar submission</a>
@@ -1276,7 +1276,7 @@
 
 {% macro acmg_form(selected=None) %}
   <label class="mt-3">ACMG classification</label>
-  <form action="{{ url_for('variants.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
+  <form action="{{ url_for('variant.variant_update', institute_id=institute._id, case_name=case.display_name, variant_id=variant._id) }}" method="POST">
     <div class="d-flex justify-content-between">
       {% for option in ACMG_OPTIONS %}
         <button class="btn btn-{{ option.color if (option.code == selected or not selected) else 'outline-secondary' }} form-control {{ 'mr-1' if not loop.last }}" name="acmg_classification" value="{{ option.code }}" title="{{ option.label }}">
@@ -1292,7 +1292,7 @@
   <li class="list-group-item {{ 'list-group-item-info' if current_variant }}">
     <div class="d-flex">
       <span>
-        <a href="{{ url_for('variants.evaluation', evaluation_id=data._id) }}">
+        <a href="{{ url_for('variant.evaluation', evaluation_id=data._id) }}">
           {{ data.classification.label }}
         </a>
         <span class="badge badge-info">{{ data.classification.short }}</span>
@@ -1304,7 +1304,7 @@
       </span>
     </div>
     <small class="text-muted">
-      <form action="{{ url_for('variants.evaluation', evaluation_id=data._id) }}" method="POST">
+      <form action="{{ url_for('variant.evaluation', evaluation_id=data._id) }}" method="POST">
       {{ data.user_name }} on {{ data.created_at.date() }}
       {% if current_variant %}
         <button class="btn btn-xs btn-link">Delete</button>
