+++ scout/server/blueprints/variants/controllers.py
@@ -151,12 +151,12 @@ def parse_variant(store, institute_obj, case_obj, variant_obj, update=False, gen
                                            variant_id=variant_obj['variant_id'], comments=True)
 
     if variant_genes:
-        variant_obj.update(get_predictions(variant_genes))
+        variant_obj.update(predictions(variant_genes))
         if variant_obj.get('category') == 'cancer':
             variant_obj.update(get_variant_info(variant_genes))
 
     for compound_obj in compounds:
-        compound_obj.update(get_predictions(compound_obj.get('genes', [])))
+        compound_obj.update(predictions(compound_obj.get('genes', [])))
 
     classification = variant_obj.get('acmg_classification')
     if isinstance(classification, int):
@@ -291,252 +291,6 @@ def get_variant_info(genes):
         data['canonical_transcripts'].append(value)
 
     return data
-
-
-def get_predictions(genes):
-    """Get sift predictions from genes."""
-    data = {
-        'sift_predictions': [],
-        'polyphen_predictions': [],
-        'region_annotations': [],
-        'functional_annotations': []
-    }
-    for gene_obj in genes:
-        for pred_key in data:
-            gene_key = pred_key[:-1]
-            if len(genes) == 1:
-                value = gene_obj.get(gene_key, '-')
-            else:
-                gene_id = gene_obj.get('hgnc_symbol') or str(gene_obj['hgnc_id'])
-                value = ':'.join([gene_id, gene_obj.get(gene_key, '-')])
-            data[pred_key].append(value)
-
-    return data
-
-def parse_gene(gene_obj, build=None):
-    """Parse variant genes."""
-    build = build or 37
-
-    if gene_obj.get('common'):
-        add_gene_links(gene_obj, build)
-        refseq_transcripts = []
-        for tx_obj in gene_obj['transcripts']:
-            parse_transcript(gene_obj, tx_obj, build)
-
-            # select refseq transcripts as "primary"
-            if not tx_obj.get('refseq_id'):
-                continue
-
-            refseq_transcripts.append(tx_obj)
-
-        gene_obj['primary_transcripts'] = (refseq_transcripts if refseq_transcripts else [])
-
-def parse_transcript(gene_obj, tx_obj, build=None):
-    """Parse variant gene transcript (VEP)."""
-    build = build or 37
-    add_tx_links(tx_obj, build)
-
-    if tx_obj.get('refseq_id'):
-        gene_name = (gene_obj['common']['hgnc_symbol'] if gene_obj['common'] else
-                     gene_obj['hgnc_id'])
-        tx_obj['change_str'] = transcript_str(tx_obj, gene_name)
-
-def transcript_str(transcript_obj, gene_name=None):
-    """Generate amino acid change as a string."""
-    if transcript_obj.get('exon'):
-        gene_part, part_count_raw = 'exon', transcript_obj['exon']
-    elif transcript_obj.get('intron'):
-        gene_part, part_count_raw = 'intron', transcript_obj['intron']
-    else:
-        # variant between genes
-        gene_part, part_count_raw = 'intergenic', '0'
-
-    part_count = part_count_raw.rpartition('/')[0]
-    change_str = "{}:{}{}:{}:{}".format(
-        transcript_obj.get('refseq_id', ''),
-        gene_part,
-        part_count,
-        transcript_obj.get('coding_sequence_name', 'NA'),
-        transcript_obj.get('protein_sequence_name', 'NA'),
-    )
-    if gene_name:
-        change_str = "{}:".format(gene_name) + change_str
-    return change_str
-
-
-def end_position(variant_obj):
-    """Calculate end position for a variant."""
-    alt_bases = len(variant_obj['alternative'])
-    num_bases = max(len(variant_obj['reference']), alt_bases)
-    return variant_obj['position'] + (num_bases - 1)
-
-
-def frequency(variant_obj):
-    """Returns a judgement on the overall frequency of the variant.
-
-    Combines multiple metrics into a single call.
-    """
-    most_common_frequency = max(variant_obj.get('thousand_genomes_frequency') or 0,
-                                variant_obj.get('exac_frequency') or 0)
-    if most_common_frequency > .05:
-        return 'common'
-    elif most_common_frequency > .01:
-        return 'uncommon'
-    else:
-        return 'rare'
-
-
-def clinsig_human(variant_obj):
-    """Convert to human readable version of CLINSIG evaluation."""
-    for clinsig_obj in variant_obj['clnsig']:
-        # The clinsig objects allways have a accession
-        if isinstance(clinsig_obj['accession'], int):
-            # New version
-            link = "https://www.ncbi.nlm.nih.gov/clinvar/variation/{}"
-        else:
-            # Old version
-            link = "https://www.ncbi.nlm.nih.gov/clinvar/{}"
-
-        human_str = 'not provided'
-        if clinsig_obj.get('value'):
-            try:
-                # Old version
-                int(clinsig_obj['value'])
-                human_str = CLINSIG_MAP.get(clinsig_obj['value'], 'not provided')
-            except ValueError:
-                # New version
-                human_str = clinsig_obj['value']
-
-        clinsig_obj['human'] = human_str
-        clinsig_obj['link'] = link.format(clinsig_obj['accession'])
-
-        yield clinsig_obj
-
-
-def thousandg_link(variant_obj, build=None):
-    """Compose link to 1000G page for detailed information."""
-    dbsnp_id = variant_obj.get('dbsnp_id')
-    build = build or 37
-
-    if not dbsnp_id:
-        return None
-
-    if build == 37:
-        url_template = ("http://grch37.ensembl.org/Homo_sapiens/Variation/Explore"
-                        "?v={};vdb=variation")
-    else:
-        url_template = ("http://www.ensembl.org/Homo_sapiens/Variation/Explore"
-                        "?v={};vdb=variation")
-
-    return url_template.format(dbsnp_id)
-
-
-def exac_link(variant_obj):
-    """Compose link to ExAC website for a variant position."""
-    url_template = ("http://exac.broadinstitute.org/variant/"
-                    "{this[chromosome]}-{this[position]}-{this[reference]}"
-                    "-{this[alternative]}")
-    return url_template.format(this=variant_obj)
-
-
-def gnomad_link(variant_obj):
-    """Compose link to gnomAD website."""
-    url_template = ("http://gnomad.broadinstitute.org/variant/{this[chromosome]}-"
-                    "{this[position]}-{this[reference]}-{this[alternative]}")
-    return url_template.format(this=variant_obj)
-
-
-def swegen_link(variant_obj):
-    """Compose link to SweGen Variant Frequency Database."""
-    url_template = ("https://swegen-exac.nbis.se/variant/{this[chromosome]}-"
-                    "{this[position]}-{this[reference]}-{this[alternative]}")
-    return url_template.format(this=variant_obj)
-
-def cosmic_link(variant_obj):
-    """Compose link to COSMIC Database.
-
-    Args:
-        variant_obj(scout.models.Variant)
-
-    Returns:
-        url_template(str): Link to COSMIIC database if cosmic id is present
-    """
-
-    cosmic_ids = variant_obj.get('cosmic_ids')
-
-    if not cosmic_ids:
-        return None
-    else:
-        cosmic_id = cosmic_ids[0]
-        url_template = ("https://cancer.sanger.ac.uk/cosmic/mutation/overview?id={}")
-
-
-    return url_template.format(cosmic_id)
-
-def beacon_link(variant_obj, build=None):
-    """Compose link to Beacon Network."""
-    build = build or 37
-    url_template = ("https://beacon-network.org/#/search?pos={this[position]}&"
-                    "chrom={this[chromosome]}&allele={this[alternative]}&"
-                    "ref={this[reference]}&rs=GRCh37")
-    # beacon does not support build 38 at the moment
-    # if build == '38':
-    #     url_template = ("https://beacon-network.org/#/search?pos={this[position]}&"
-    #                     "chrom={this[chromosome]}&allele={this[alternative]}&"
-    #                     "ref={this[reference]}&rs=GRCh38")
-
-    return url_template.format(this=variant_obj)
-
-
-def ucsc_link(variant_obj, build=None):
-    """Compose link to UCSC."""
-    build = build or 37
-    url_template = ("http://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&"
-                        "position=chr{this[chromosome]}:{this[position]}"
-                        "-{this[position]}&dgv=pack&knownGene=pack&omimGene=pack")
-    if build == 38:
-        url_template = ("http://genome.ucsc.edu/cgi-bin/hgTracks?db=hg20&"
-                        "position=chr{this[chromosome]}:{this[position]}"
-                        "-{this[position]}&dgv=pack&knownGene=pack&omimGene=pack")
-
-    return url_template.format(this=variant_obj)
-
-
-def alamut_link(variant_obj):
-    url_template = ("http://localhost:10000/show?request={this[chromosome]}:"
-                    "{this[position]}{this[reference]}>{this[alternative]}")
-    return url_template.format(this=variant_obj)
-
-
-def spidex_human(variant_obj):
-    """Translate SPIDEX annotation to human readable string."""
-    if variant_obj.get('spidex') is None:
-        return 'not_reported'
-    elif abs(variant_obj['spidex']) < SPIDEX_HUMAN['low']['pos'][1]:
-        return 'low'
-    elif abs(variant_obj['spidex']) < SPIDEX_HUMAN['medium']['pos'][1]:
-        return 'medium'
-    else:
-        return 'high'
-
-
-def expected_inheritance(variant_obj):
-    """Gather information from common gene information."""
-    manual_models = set()
-    for gene in variant_obj.get('genes', []):
-        manual_models.update(gene.get('manual_inheritance', []))
-    return list(manual_models)
-
-
-def callers(variant_obj, category='snv'):
-    """Return info about callers."""
-    calls = set()
-    for caller in CALLERS[category]:
-        if variant_obj.get(caller['id']):
-            calls.add((caller['name'], variant_obj[caller['id']]))
-
-    return list(calls)
-
 
 def variant_verification(store, mail, institute_obj, case_obj, user_obj, variant_obj, sender, variant_url, order, comment, url_builder=url_for):
     """Sand a verification email and register the verification in the database
@@ -743,40 +497,6 @@ def cancer_variants(store, institute_id, case_name, form, page=1):
     )
     return data
 
-def clinvar_export(store, institute_id, case_name, variant_id):
-    """Gather the required data for creating the clinvar submission form
-
-        Args:
-            store(scout.adapter.MongoAdapter)
-            institute_id(str): Institute ID
-            case_name(str): case ID
-            variant_id(str): variant._id
-
-        Returns:
-            a dictionary with all the required data (case and variant level) to pre-fill in fields in the clinvar submission form
-
-    """
-
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    pinned = [store.variant(variant_id) or variant_id for variant_id in
-                  case_obj.get('suspects', [])]
-    variant_obj = store.variant(variant_id)
-
-    # gather missing transcript info from entrez (refseq id version)
-    for pinned_var in pinned:
-        for gene in pinned_var.get('genes'):
-            for transcript in gene.get('transcripts'):
-                if transcript.get('refseq_id'):
-                    transcript['refseq_id'] = fetch_refseq_version(transcript['refseq_id'])
-
-    return dict(
-        today = str(date.today()),
-        institute=institute_obj,
-        case=case_obj,
-        variant=variant_obj,
-        pinned_vars=pinned
-    )
-
 def get_clinvar_submission(store, institute_id, case_name, variant_id, submission_id):
     """Collects all variants from the clinvar submission collection with a specific submission_id
 
@@ -804,44 +524,6 @@ def get_clinvar_submission(store, institute_id, case_name, variant_id, submissio
         pinned_vars=pinned,
         clinvars = clinvar_submission_objs
     )
-
-
-def variant_acmg(store, institute_id, case_name, variant_id):
-    """Collect data relevant for rendering ACMG classification form."""
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj = store.variant(variant_id)
-    return dict(institute=institute_obj, case=case_obj, variant=variant_obj,
-                CRITERIA=ACMG_CRITERIA, ACMG_OPTIONS=ACMG_OPTIONS)
-
-
-def variant_acmg_post(store, institute_id, case_name, variant_id, user_email, criteria):
-    """Calculate an ACMG classification based on a list of criteria."""
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj = store.variant(variant_id)
-    user_obj = store.user(user_email)
-    variant_link = url_for('variant.variant', institute_id=institute_id,
-                           case_name=case_name, variant_id=variant_id)
-    classification = store.submit_evaluation(
-        institute_obj=institute_obj,
-        case_obj=case_obj,
-        variant_obj=variant_obj,
-        user_obj=user_obj,
-        link=variant_link,
-        criteria=criteria,
-    )
-    return classification
-
-
-def evaluation(store, evaluation_obj):
-    """Fetch and fill-in evaluation object."""
-    evaluation_obj['institute'] = store.institute(evaluation_obj['institute_id'])
-    evaluation_obj['case'] = store.case(evaluation_obj['case_id'])
-    evaluation_obj['variant'] = store.variant(evaluation_obj['variant_specific'])
-    evaluation_obj['criteria'] = {criterion['term']: criterion for criterion in
-                                  evaluation_obj['criteria']}
-    evaluation_obj['classification'] = ACMG_COMPLETE_MAP.get(evaluation_obj['classification'])
-    return evaluation_obj
-
 
 def upload_panel(store, institute_id, case_name, stream):
     """Parse out HGNC symbols from a stream."""
